% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_est.R
\name{calcular_est}
\alias{calcular_est}
\title{Create standard columns and types from data in BITS shapefiles}
\usage{
calcular_est(estados, edo_val, bits_df, marco_muestral)
}
\arguments{
\item{estados}{string ot vector of strings with the name of the state as
defined in sampling frame}

\item{edo_val}{data.frame with expert classification.}

\item{bits_df}{data.frame with BITS classifications.}
}
\value{
A list with estimates for percentage of area correcrly classified and
  percentage of polygons correctly classified.
}
\description{
Data in different tiles has diferent columns, this function standardizes the
information and variable types.
}
\examples{
sampling_frame <- data.frame(id = 1:100,
  str = sample(1:5, 100, replace = TRUE),
  val = rnorm(100))
allo <- sampling_frame \%>\%
    group_by(str) \%>\%
    summarise(n = 0.4 * n())
select_sample(allo, sampling_frame, n, str)
}
