% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_ratio_est.R
\name{combined_ratio_est}
\alias{combined_ratio_est}
\title{Compute ratio estimator for strata and combined ratio estimator for a
stratified sample of polygons.}
\usage{
combined_ratio_est(data, allocation, stratum, M, y, N)
}
\arguments{
\item{data}{\code{data.frame} each line is an observation and includes the
stratum and columns used in ratio estimation.}

\item{allocation}{\code{data.frame} with a column defining the strata and a
column with sample size allocations for each stratum (one line per stratum).}

\item{stratum}{unquoted column with strata in the data and allocation
\code{data.frame}'s (the columns must have the same name in the
two \code{data.frame}'s)}

\item{M:}{unquoted variable in data corresponding to denominator in ratio
(eg. polygon areas)}

\item{y:}{unquoted variable in data corresponding to numerator in ratio
(eg. binary vector 1 correctly classified 0 ow)}

\item{N:}{number of polygons in the strata.}
}
\value{
A \code{list} with estimates per strata and the combined ratio
  estimate for overall population.
}
\description{
Ratio estimator is computed for each strata and then a combined version of
the ratio estimator is computed where the estimates of the strata are
weighted according to the area of each stratum (unlike traditional estimator
that weights with number of polygons sampled).
\code{strata_ratio_est} returns the estimates for each strata in the sample
 (assuming SRS within strata).
}
\examples{
sampling_frame <- data_frame(id = 1:100, str = sample(1:5, 100, replace = TRUE))
allo <- sampling_frame \%>\%
    group_by(str) \%>\%
    mutate(n = 0.4 * n())
select_sample(allo, sampling_frame, n, str)
}
