% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sample.R
\name{select_sample}
\alias{select_sample}
\title{Randomly select samples from a sampling frame with or without strata.}
\usage{
select_sample(allocation, sampling_frame, sample_size = sample_size,
  stratum = stratum)
}
\arguments{
\item{allocation}{\code{data.frame} with a column defining the strata and a
column with sample size allocations for each stratum (one line per stratum).}

\item{sampling_frame}{\code{data.frame} with the sampling frame it must contain a
column with the stratum.}

\item{sample_size}{unquoted column with sample sizes in the allocation
data.frame}

\item{stratum}{unquoted column with strata in the allocation and
sampling_frame \code{data.frame}'s (the columns must have the same name in the
two \code{data.frame}'s)}
}
\value{
A \code{data.frame} with the selected sample, it will have the same
  columns as the original sampling frame plus a column indicating the sample
  size in the stratum of each selected observation.
}
\description{
Given a data.frame with sample size allocation per strata and a data.frame
with the sampling frame it selects random samples.
}
\examples{
sampling_frame <- data.frame(id = 1:100,
  str = sample(1:5, 100, replace = TRUE),
  val = rnorm(100))
allo <- sampling_frame \%>\%
    group_by(str) \%>\%
    summarise(n = 0.4 * n())
select_sample(allo, sampling_frame, n, str)
}
