% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_ratio_est.R
\name{separate_ratio_est}
\alias{separate_ratio_est}
\title{Compute ratio estimator for strata and separate ratio estimator for a
stratified sample of polygons.}
\usage{
separate_ratio_est(data, estratos_area, stratum, y, M, N)
}
\arguments{
\item{data}{\code{data.frame} each line is an observation and includes the
stratum and columns used in ratio estimation.}

\item{stratum}{unquoted column with strata in the data and allocation
\code{data.frame}'s (the columns must have the same name in the
two \code{data.frame}'s)}

\item{allocation}{\code{data.frame} with a column defining the strata and a
column with sample size allocations for each stratum (one line per stratum).}

\item{y:}{unquoted variable in data corresponding to numerator in ratio
(eg. binary vector 1 correctly classified 0 ow)
#' @param M: unquoted variable in data corresponding to denominator in ratio
(eg. polygon areas)}

\item{N:}{number of polygons in the strata.}
}
\value{
A \code{list} with estimates per strata and the separate ratio
  estimate for overall population.
}
\description{
Ratio estimator is computed for each strata and then a separate version of
the ratio estimator is computed where the estimates of the strata are
weighted according to the area of each stratum (unlike traditional estimator
that weights with number of polygons sampled).
\code{strata_ratio_est} returns the estimates for each strata in the sample
 (assuming SRS within strata).
}
